requirements.txt
Flask
Pillow
torch==2.5.1
torchvision